requirements.txt
aiogram
python-dotenv
fastapi
uvicorn